// lib/enums/eApgWglTestMaterialTypes.ts
export enum eApgWglTestMaterialsTypes {
  LineMaterials = 'LineMats',

  BasicWireframeMaterials = 'Basic wireframe',

  LambertMaterials = 'Lambert',
  LambertFlatMaterials = 'Lambert Flat',
  LambertFlatTransparentMaterials = 'Lambert Flat Transparent',

  PhongMaterials = 'Phong',
  PhongTexturizedMaterials = 'Phong Texturized',
  PhongBumpMappedMaterials = 'Phong Bump Mapped',
  PhongNormalMappedMaterials = 'Phong Normal Mapped',

  StdMaterials = 'Standard',
  StdTexturizedMaterials = 'Standard Texturized',
  StdBumpMappedMaterials = 'Standard Bump Mapped',
  StdNormalMappedMaterials = 'Standard Normal Mapped'
}